Changes to allow packaging for local dev
