+++ requirements/base.pip
@@ -1,0 +1,7 @@
+wheel
+xlrd
+openpyxl
+pandas
+pyomo
+matplotlib
+
