+++ solver_links.txt
@@ -1,0 +1,1 @@
+For Mac : https://formulae.brew.sh/formula/glpk
